requirements.txt
streamlit==1.52.2
openai>=1.0.0
python-dotenv>=0.21.0
requests>=2.28.0
pandas>=1.5.0
numpy>=1.23.0
pillow>=9.0.0